fix: vadjust llm monkey patch for vllm>=0.2.7
