+++ openrlhf/trainer/ray/__init__.py
@@ -1,3 +1,4 @@
 from .launcher import DistributedTorchRayActor, PPORayActorGroup, ReferenceModelRayActor, RewardModelRayActor
 from .ppo_actor import ActorModelRayActor
 from .ppo_critic import CriticModelRayActor
+from .vllm_engine import create_vllm_engines
